Labels underneath map (and a bit of timer) works
There is now a getMapController() method in MasterController that will
get our controller for Map. This allows us to call methods like
setPhaseLabel() and startTimer().

Game now has a startTurn() method that performs all of the logic
necessary at the beginning of a turn (right now setting labels and
timer)